Implemented Switching of Trigram Read Ordering
While likely not going to be useful, I implemented the ability to change the readmode of the trigram order.

This is also designed to allow different read modes to be viewed at the same time without causing 'bleedover' of the readmode when updating other message panes.

Seems like it'd be simple but we I needed to implement a host of functionality to keep track of which pane had what selected last. (if it fails to find the latest, it'll use the default readmode)